requirements.txt
boto3>=1.9.26
botocore>=1.12.26
cairocffi>=0.9.0
CairoSVG>=1.0.22
certifi>=2018.10.15
cffi>=1.11.5
chardet>=3.0.4
contextlib2>=0.5.5
cssselect2>=0.2.1
decorator>=4.3.0
diff-match-patch>=20121119
Django>=1.8.19,<2.1
django-appconf>=1.0.2
django-compressor>=2.2
django-constance>=2.3.1
django-extensions>=2.1.3
django-filer>=1.2.4
django-import-export>=1.1.0
django-mptt>=0.8.7
django-multisite>=1.6.0
django-polymorphic>=0.8.1,<2.1
django-request>=1.5.4
django-reversion>=3.0.0
django-settings-export>=1.2.1
django-sitetree>=1.12.0
django-storages>=1.6.5
django-suit>=0.2.19
django-user-agents>=0.3.2
django-xforwardedfor-middleware>=1.1
djangorestframework>=3.6.4
docutils>=0.14
easy-thumbnails>=2.3
et-xmlfile>=1.0.1
futures>=3.1.0
gevent>=1.3.7
greenlet>=0.4.15
gunicorn>=19.9.0
html5lib>=1.0.1
httplib2>=0.11.3
huey>=1.2.1
idna>=2.7
jdcal>=1.4
jmespath>=0.9.3
jsonfield>=2.0.2
lxml>=4.2.5
mistune>=0.8.4
odfpy>=1.3.6
openpyxl>=2.5.8
pdfrw>=0.4
Pillow>=5.3.0
plivo>=4.1.3
psycopg2-binary>=2.7.5
pycparser>=2.19
pyparsing>=2.2.2
Pyphen>=0.9.5
python-dateutil>=2.7.3,<2.8.1
pytz>=2018.5
PyYAML>=3.13
raven>=6.9.0
rcssmin>=1.0.6
redis>=2.10.6
requests>=2.19.1
requests-file>=1.4.3
rjsmin>=1.0.12
rules>=2.0
s3transfer>=0.1.13
six>=1.11.0
tablib>=0.12.1
tinycss2>=0.6.1
tldextract>=2.2.0
twilio>=5.7.0
typing>=3.6.6
ua-parser>=0.8.0
unicodecsv>=0.14.1
Unidecode>=0.4.21,<1.1
urllib3>=1.23
user-agents>=1.1.0
virtualenv>=16.0.0
WeasyPrint>=0.42.3
webencodings>=0.5.1
whitenoise>=4.1
xlrd>=1.1.0
xlwt>=1.3.0
django-npm>=1.0.0
backports.csv
future>=0.18.2
